Create not transport specific request abstractions
They're specialised on one kind of request. I intend to use these to be able to
provide useful information in structured exceptions generated on the transport
layer.

Before this, information about what kind of request we're processing, and
therefore what different HTTP status codes meant in the context of that request,
were already lost at that level.