Sleep after issuing DELETEs, not before
The other way around, list_keys might return keys that have already been asked
to be deleted, but who's deletion hasn't propagated to all the replicas yet,
making them show up in list_keys again.

The second DELETE for a key would also sometimes cause an exception due to a
missing vclock header in the response expected. This is a separate issue that
needs to be looked into.